created sign in/out component and added alies
