+++ package.json
@@ -3,6 +3,9 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
+  "browser": {
+    "[module-name]": false
+  },
   "scripts": {
     "start": "cross-env SERVE=true webpack serve --open --port 3000",
     "build": "webpack",
@@ -50,6 +53,7 @@
     "bootstrap-icons": "^1.10.3",
     "concurrently": "^7.6.0",
     "ky": "^0.33.2",
+    "module-alias": "^2.2.2",
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.2",
     "react-dom": "^18.2.0",
