+++ src/App.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 import { BrowserRouter, Route, Routes, Navigate } from "react-router-dom";
-import NavBar from "./components/Navbar";
-import RegistrationPage from "./pages/RegistrationPage";
-import MainPage from "./pages/MainPage";
-import LoginPage from "./pages/LoginPage";
-import MyPage from "./pages/MyPage";
-import { useTypedSelector } from "./hooks/useTypedSelector";
+import NavBar from "@/components/Navbar";
+import RegistrationPage from "@/pages/RegistrationPage";
+import MainPage from "@/pages/MainPage";
+import LoginPage from "@/pages/LoginPage";
+import MyPage from "@/pages/MyPage";
+import { useTypedSelector } from "@/hooks/useTypedSelector";
 
 const App = () => {
   const { isLogin } = useTypedSelector((state) => state);
