+++ src/components/Navbar.tsx
@@ -1,22 +1,12 @@
 import React, { FC } from "react";
 import { NavLink, useNavigate } from "react-router-dom";
 import { Container, Nav, Navbar } from "react-bootstrap";
-import "../styles/Navbar.scss";
-import { useTypedSelector } from "../hooks/useTypedSelector";
-const logo = require("../images/logo.png");
+import "@/styles/Navbar.scss";
+import SignInOut from "./SignInOut";
+
+const logo = require("@/images/logo.png");
 
 const NavBar: FC = () => {
-  const { isLogin } = useTypedSelector((store) => store);
-  const navigate = useNavigate();
-  const goToLink = (event: React.MouseEvent<HTMLAnchorElement>): void => {
-    event.preventDefault();
-    if (isLogin) {
-      sessionStorage.removeItem("token");
-      window.open("http://localhost:5000/api/users/logout", "_self");
-    } else {
-      navigate("/login");
-    }
-  };
   return (
     <Navbar bg="dark" variant="dark">
       <Container>
@@ -35,13 +25,7 @@ const NavBar: FC = () => {
           <NavLink to={"/mypage"} className="nav-link">
             <i className="bi bi-person-circle"></i>
           </NavLink>
-          <NavLink to={"/login"} className="nav-link" onClick={goToLink}>
-            {isLogin ? (
-              <i className="bi bi-box-arrow-in-left"></i>
-            ) : (
-              <i className="bi bi-box-arrow-in-right"></i>
-            )}
-          </NavLink>
+          <SignInOut />
         </Nav>
       </Container>
     </Navbar>
