+++ src/hooks/useTypedSelector.ts
@@ -1,4 +1,4 @@
 import { TypedUseSelectorHook, useSelector } from "react-redux";
-import { StateRedux } from "../store";
+import { StateRedux } from "@/store";
 
 export const useTypedSelector: TypedUseSelectorHook<StateRedux> = useSelector;
