+++ src/pages/RegistrationPage.tsx
@@ -7,7 +7,7 @@ import * as Yup from "yup";
 import ky from "ky";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
-import "../styles/Auth.scss";
+import "@/styles/Auth.scss";
 
 type UserRegistrationForm = {
   username: string;
