+++ webpack.config.js
@@ -1,6 +1,7 @@
 const path = require("path");
 const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 const ReactRefreshWebpackPlugin = require("@pmmmwh/react-refresh-webpack-plugin");
+const TsconfigPathsPlugin = require("tsconfig-paths-webpack-plugin");
 
 let mode = "development";
 if (process.env.NODE_ENV === "production") {
@@ -11,6 +12,7 @@ const plugins = [
   new MiniCssExtractPlugin({
     filename: "[name].css",
   }),
+  new TsconfigPathsPlugin(),
 ];
 
 if (process.env.SERVE) {
@@ -30,6 +32,9 @@ module.exports = {
   },
   resolve: {
     extensions: [".ts", ".tsx", ".js"],
+    alias: {
+      "@": path.resolve(__dirname, "src"),
+    },
   },
   devServer: {
     hot: true,
